requirements.txt
anyio==2.2.0
asks==2.4.12
async-generator==1.10
attrs==20.3.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
filelock==3.0.12
gdown==3.12.2
h11==0.12.0
idna==2.10
lxml==4.6.3
mega.py==1.0.8
outcome==1.1.0
pathlib==1.0.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2021.1
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.3.0
soupsieve==2.2.1
tenacity==5.1.5
tqdm==4.60.0
trio==0.18.0
urllib3==1.26.4
wget==3.2
